\name{Eval.criteria}
\alias{Eval.criteria}
\alias{Eval.seq.criteria}
\alias{ExpT}
\alias{FImatrix}
\alias{gen.data.EC}
\alias{ind.FIM}
\alias{likelihood}
\alias{MCMC.draw.EC}
\alias{mnius.loglikelihood}
\alias{mu.ders}
\alias{next.point}
\alias{ori.level}
\alias{StandLevel}
\alias{StandStressFun}
\alias{TMCMC}
\alias{TMCMCALT}
\alias{trad.design}
\alias{use.profile}
\alias{use}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal functions
}
\description{Internal functions
}
\keyword{internal
}